refactor: move the openapi not found exception to exceptions module
